Refactor
Refactored folder structure.
Updated init and hook files to account for new implementation of require.
Removed import function (replaced with require).
