+++ LuaEx/class/stack.lua
