+++ LuaEx/lib/constant.lua
