+++ LuaEx/lib/enum.lua
